feat(nix): add systemd user service module
This module can be used to set up the cli automatically upon a
graphical-session starting.

The way this is written only allows for one instance of the CLI to be
running. The module will either need to be changed or the CLI adapted to
accept multiple arguments if this feature is desired.

The `display` option will set the `DISPLAY` environment variable for the
program, which will be used by libxdo to route presses. In my testing
this needs to be set to `:1` when running under wayland and `:0` when
running under X. If you know what you are doing these values can be
different. I've defaulted it to `null` so that no environment variable
will be set if it is not included.